# src/openapi/system-spec.yaml
openapi: 3.0.3
info:
  title: HAXcms System API
  description: Control-plane API for authentication, lifecycle/admin, discovery/meta, and system settings/configuration.
  version: 0.0.0
servers:
  - url: /
    description: System API base path
...
paths:
...
  /system/api/v1/system/app-store:
    get:
      tags:
        - settings
      operationId: generateAppStore
      summary: Generate app-store provider manifest for the editor
      security:
        - appStoreTokenQuery: []
          siteTokenQuery: []
      responses:
        "200":
          description: App-store payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /system/api/v1/settings/api-keys:
    get:
      tags:
        - settings
      operationId: getApiKeys
      summary: Return configured API keys and provider statuses
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: API key settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - settings
      operationId: saveApiKeysPost
      summary: Update API key settings
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeysSettings"
      responses:
        "200":
          description: API key update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - settings
      operationId: saveApiKeysPatch
      summary: Update API key settings
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeysSettings"
      responses:
        "200":
          description: API key update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/settings/media:
    get:
      tags:
        - settings
      operationId: getMediaSettings
      summary: Return media and upload configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Media settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - settings
      operationId: saveMediaSettingsPost
      summary: Update media and upload configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaSettings"
      responses:
        "200":
          description: Media settings update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - settings
      operationId: saveMediaSettingsPatch
      summary: Update media and upload configuration
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaSettings"
      responses:
        "200":
          description: Media settings update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /system/api/v1/settings/schema-files/operations:
    post:
      tags:
        - settings
      operationId: schemaFileOperation
      summary: Perform schema file operation in system configuration storage
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Schema file operation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
...
  responses:
...
    Forbidden:
      description: Authenticated user is not authorized for this route
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    userTokenQuery:
      type: apiKey
      in: query
      name: user_token
    userTokenHeader:
      type: apiKey
      in: header
      name: X-HAXCMS-User-Token
    siteTokenQuery:
      type: apiKey
      in: query
      name: site_token
    appStoreTokenQuery:
      type: apiKey
      in: query
      name: appstore_token
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status:
          type: integer
        data: {}
      additionalProperties: true
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            message:
              type: string
...
    ApiKeysSettings:
      type: object
      description: API key settings payload
      additionalProperties: true
    MediaSettings:
      type: object
      description: Media settings payload
      additionalProperties: true
...